# client/src/openapi/openapi.json
{
    "openapi": "3.0.1",
    "info": {
        "title": "OpenApi specification - Portal Microservice",
        "description": "OpenApi documentation for Portal Microservice",
        "contact": {
            "name": "Miguel Cardoso",
            "url": "https://www.linkedin.com/in/miguelcardoso19/",
            "email": "miguel40cardoso@gmail.com"
        },
        "version": "1.0"
    },
    "servers": [
        {
            "url": "http://localhost:8080",
            "description": "Local env"
        }
    ],
...
    "paths": {
...
        "/api/v1/invoice/update": {
            "put": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Update an invoice",
                "description": "Updates an existing invoice record.",
                "operationId": "updateInvoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "400": {
                        "description": "Invalid invoice data",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Invoice not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Invoice updated successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/invoice/create": {
            "post": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Create a new invoice",
                "description": "Creates a new invoice record.",
                "operationId": "createInvoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/InvoiceDTO"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Invoice created successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid invoice data",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/invoice/attach-multipart-file/{invoiceId}": {
            "post": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Attach multipart file to invoice",
                "description": "Attaches a file to an existing invoice using Feign client.",
                "operationId": "attachMultipartFileToInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Failed to upload file"
                    },
                    "404": {
                        "description": "Invoice not found"
                    },
                    "204": {
                        "description": "File successfully uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/invoice/attach-base64-file/{invoiceId}": {
            "post": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Attach base64 file to invoice",
                "description": "Attaches a base64 encoded file to an invoice.",
                "operationId": "attachBase64FileToInvoice",
                "parameters": [
                    {
                        "name": "invoiceId",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "file"
                                ],
                                "type": "object",
                                "properties": {
                                    "file": {
                                        "type": "string",
                                        "format": "binary"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "400": {
                        "description": "Failed to upload file"
                    },
                    "204": {
                        "description": "File successfully uploaded"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/v1/invoice/{id}": {
            "get": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Get an invoice by ID",
                "description": "Fetches an invoice by its unique identifier.",
                "operationId": "getInvoiceById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Invoice not found",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Invoice fetched successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/InvoiceDTO"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Delete an invoice",
                "description": "Deletes an invoice by its ID.",
                "operationId": "deleteInvoice",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "404": {
                        "description": "Invoice not found"
                    },
                    "204": {
                        "description": "Invoice deleted successfully"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/v1/invoice/all": {
            "get": {
                "tags": [
                    "Invoice Controller"
                ],
                "summary": "Get all invoices",
                "description": "Fetches a paginated list of all invoices.",
                "operationId": "getAllInvoices",
                "parameters": [
                    {
                        "name": "pageable",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "$ref": "#/components/schemas/Pageable"
                        }
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid pagination parameters",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPageDTOObject"
                                }
                            }
                        }
                    },
                    "200": {
                        "description": "Invoices fetched successfully",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPageDTOObject"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "*/*": {
                                "schema": {
                                    "$ref": "#/components/schemas/CustomPageDTOObject"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "InvoiceDTO": {
                "required": [
                    "correlationId",
                    "dateOfEmission",
                    "description",
                    "id",
                    "version"
                ],
                "type": "object",
                "properties": {
                    "correlationId": {
                        "type": "string",
                        "description": "Unique correlation ID to trace and associate requests/responses.",
                        "format": "uuid",
                        "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                    },
                    "id": {
                        "type": "string",
                        "description": "Unique identifier for the entity, typically assigned upon creation or retrieval.",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "version": {
                        "type": "integer",
                        "description": "Version number for optimistic locking and concurrent updates management.",
                        "format": "int32",
                        "example": 1
                    },
                    "dateOfEmission": {
                        "type": "string",
                        "description": "The date when the invoice was emitted.",
                        "format": "date",
                        "example": "2024-11-10"
                    },
                    "description": {
                        "type": "string",
                        "description": "A brief description of the invoice.",
                        "example": "Invoice for marketing services"
                    },
                    "movementId": {
                        "type": "string",
                        "description": "UUID of the associated movement, if any.",
                        "format": "uuid",
                        "nullable": true,
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "file": {
                        "type": "array",
                        "description": "The file content of the invoice, stored as a byte array.",
                        "nullable": true,
                        "items": {
                            "type": "string",
                            "description": "The file content of the invoice, stored as a byte array.",
                            "format": "byte",
                            "nullable": true
                        }
                    },
                    "movement": {
                        "$ref": "#/components/schemas/MovementDTO"
                    },
                    "stripeReceiptUrl": {
                        "type": "string",
                        "description": "The URL to the receipt generated by Stripe for the transaction",
                        "nullable": true,
                        "example": "https://pay.stripe.com/receipts/acct_1234/ch_1JN3zd2eZvKYlo2C0Yabc123/rcpt_12345"
                    },
                    "movementDocumentNumber": {
                        "type": "string",
                        "description": "The unique document number associated with the payment movement",
                        "nullable": true,
                        "example": "INV-2024-0001"
                    },
                    "fileKey": {
                        "type": "string",
                        "description": "The key for accessing the invoice file stored in S3. This is used to retrieve or delete the file associated with the invoice.",
                        "nullable": true,
                        "example": "invoice-1234567890-xyz.pdf"
                    },
                    "fileBase64": {
                        "type": "string",
                        "description": "Base64 encoded content of the invoice file (e.g., PDF), included in the response to display the file.",
                        "nullable": true
                    }
                },
                "description": "DTO representing an invoice, including details such as the date of emission, description, associated movement, and file data.",
                "nullable": true
            },
            "MovementDTO": {
                "required": [
                    "correlationId",
                    "dateOfEmission",
                    "description",
                    "id",
                    "status",
                    "supplierId",
                    "type",
                    "valueWithoutIva",
                    "version"
                ],
                "type": "object",
                "properties": {
                    "correlationId": {
                        "type": "string",
                        "description": "Unique correlation ID to trace and associate requests/responses.",
                        "format": "uuid",
                        "example": "f47ac10b-58cc-4372-a567-0e02b2c3d479"
                    },
                    "id": {
                        "type": "string",
                        "description": "Unique identifier for the entity, typically assigned upon creation or retrieval.",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "version": {
                        "type": "integer",
                        "description": "Version number for optimistic locking and concurrent updates management.",
                        "format": "int32",
                        "example": 1
                    },
                    "supplierId": {
                        "type": "string",
                        "description": "UUID of the supplier associated with the movement.",
                        "format": "uuid",
                        "example": "123e4567-e89b-12d3-a456-426614174000"
                    },
                    "budgetTypeId": {
                        "type": "string",
                        "description": "UUID of the associated budget type, if applicable.",
                        "format": "uuid",
                        "nullable": true,
                        "example": "987e6543-e89b-12d3-a456-426614174111"
                    },
                    "budgetSubtypeId": {
                        "type": "string",
                        "description": "UUID of the associated budget subtype, if applicable.",
                        "format": "uuid",
                        "nullable": true,
                        "example": "987e6543-e89b-12d3-a456-426614174222"
                    },
                    "invoiceId": {
                        "type": "string",
                        "description": "UUID of the associated invoice for the movement, if applicable.",
                        "format": "uuid",
                        "nullable": true,
                        "example": "a1f8b87f-e8c9-4702-92d4-c60ff15c99ab"
                    },
                    "documentNumber": {
                        "type": "string",
                        "description": "Document number associated with the movement.",
                        "nullable": true,
                        "example": "INV-2024-0001"
                    },
                    "type": {
                        "type": "string",
                        "description": "Type of the movement (e.g., INCOME, EXPENSE).",
                        "example": "EXPENSE",
                        "enum": [
                            "DEBIT",
                            "CREDIT"
                        ]
                    },
                    "dateOfEmission": {
                        "type": "string",
                        "description": "Date of emission for the movement.",
                        "format": "date",
                        "example": "2024-11-10"
                    },
                    "description": {
                        "type": "string",
                        "description": "Description of the movement.",
                        "example": "Payment for marketing services"
                    },
                    "valueWithoutIva": {
                        "type": "number",
                        "description": "Value of the movement excluding VAT (IVA).",
                        "format": "double",
                        "example": 1000.0
                    },
                    "ivaRate": {
                        "type": "number",
                        "description": "IVA (VAT) rate applied to the movement, default is 0.0 if not specified.",
                        "format": "double",
                        "nullable": true,
                        "example": 0.21
                    },
                    "ivaValue": {
                        "type": "number",
                        "description": "Calculated IVA (VAT) value for the movement, based on the IVA rate and value.",
                        "format": "double",
                        "nullable": true,
                        "example": 210.0
                    },
                    "totalValue": {
                        "type": "number",
                        "description": "Total value of the movement, including IVA (VAT).",
                        "format": "double",
                        "nullable": true,
                        "example": 1210.0
                    },
                    "invoice": {
                        "$ref": "#/components/schemas/InvoiceDTO"
                    },
                    "budgetType": {
                        "$ref": "#/components/schemas/BudgetTypeDTO"
                    },
                    "budgetSubtype": {
                        "$ref": "#/components/schemas/BudgetSubtypeDTO"
                    },
                    "supplier": {
                        "$ref": "#/components/schemas/SupplierDTO"
                    },
                    "status": {
                        "type": "string",
                        "description": "Status of the movement, such as ACCEPTED, PENDING, etc. Default is 'ACCEPTED'.",
                        "example": "ACCEPTED",
                        "enum": [
                            "ACCEPTED",
                            "PROCESSING",
                            "REFUSED",
                            "CANCELED",
                            "FAILED",
                            "REFUNDED",
                            "SUCCEEDED"
                        ]
                    }
                },
                "description": "DTO representing a financial movement, including details such as the supplier, type, value, associated invoice, and budget information."
            },
...
            "Pageable": {
                "type": "object",
                "properties": {
                    "page": {
                        "minimum": 0,
                        "type": "integer",
                        "format": "int32"
                    },
                    "size": {
                        "minimum": 1,
                        "type": "integer",
                        "format": "int32"
                    },
                    "sort": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            },
            "CustomPageDTOObject": {
                "required": [
                    "content",
                    "empty",
                    "first",
                    "last",
                    "numberOfElements",
                    "pageSize",
                    "pageable",
                    "totalElements",
                    "totalPages"
                ],
                "type": "object",
                "properties": {
                    "content": {
                        "type": "array",
                        "description": "List of content items for the current page.",
                        "items": {
                            "type": "object",
                            "description": "List of content items for the current page."
                        }
                    },
                    "pageSize": {
                        "type": "integer",
                        "description": "The number of items per page.",
                        "format": "int32",
                        "example": 20
                    },
                    "totalElements": {
                        "type": "integer",
                        "description": "Total number of elements in the entire data set.",
                        "format": "int64",
                        "example": 1000
                    },
                    "totalPages": {
                        "type": "integer",
                        "description": "Total number of pages in the data set.",
                        "format": "int32",
                        "example": 50
                    },
                    "pageable": {
                        "$ref": "#/components/schemas/CustomPageableDTO"
                    },
                    "last": {
                        "type": "boolean",
                        "description": "Indicates whether the current page is the last page.",
                        "example": false
                    },
                    "first": {
                        "type": "boolean",
                        "description": "Indicates whether the current page is the first page.",
                        "example": true
                    },
                    "numberOfElements": {
                        "type": "integer",
                        "description": "The number of elements in the current page.",
                        "format": "int32",
                        "example": 20
                    },
                    "empty": {
                        "type": "boolean",
                        "description": "Indicates whether the current page contains any content.",
                        "example": false
                    }
                },
                "description": "DTO representing a paginated response, containing a list of content items along with pagination details."
            },
...
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Insert the JWT bearer token",
                "in": "header",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    }
}